♻️ refactor utils structure, closes #42
