+++ fotoobo/utils/__init__.py
@@ -6,20 +6,15 @@ relevant fotoobo and Fortinet modules. By default the utils functions are invoke
 but they may also be accessed directly.
 """
 
-from . import ems, faz, fgt, fmg
-from .convert import convert_checkpoint
-from .fgt.config.check import fgt_config_check
-from .get import commands, inventory, version
+from . import convert, ems, faz, fgt, fmg, get
 from .greet import greet
 
 __all__ = [
-    "convert_checkpoint",
+    "convert",
     "ems",
     "faz",
     "fgt",
-    "fgt_config_check",
     "fmg",
-    "inventory",
+    "get",
     "greet",
-    "version",
 ]
