+++ fotoobo/utils/ems/__init__.py
@@ -2,8 +2,6 @@
 utils ems
 """
 
-from fotoobo.utils.ems import monitor
-
-from . import get
+from fotoobo.utils.ems import get, monitor
 
 __all__ = ["get", "monitor"]
