--- fotoobo/utils/ems/get/__init__.py
@@ -1,8 +1,0 @@
-"""
-utils ems get
-"""
-
-from .version_ import version
-from .workgroups_ import workgroups
-
-__all__ = ["version", "workgroups"]
