+++ fotoobo/utils/faz/get.py
