--- fotoobo/utils/faz/get/__init__.py
@@ -1,7 +1,0 @@
-"""
-utils faz get
-"""
-
-from .version_ import version
-
-__all__ = ["version"]
