+++ fotoobo/utils/fgt/__init__.py
@@ -3,6 +3,6 @@ utils fgt
 """
 
 from . import check, config, get
-from .backup_ import backup
+from .main import backup
 
 __all__ = ["backup", "check", "config", "get"]
