--- fotoobo/utils/fgt/check/__init__.py
@@ -1,7 +1,0 @@
-"""
-utils fgt check
-"""
-
-from .hamaster_ import hamaster
-
-__all__ = ["hamaster"]
