--- fotoobo/utils/fgt/config/__init__.py
@@ -1,7 +1,0 @@
-"""
-utils fgt config
-"""
-
-from .check import fgt_config_check
-
-__all__ = ["fgt_config_check"]
