--- fotoobo/utils/fgt/get/__init__.py
@@ -1,7 +1,0 @@
-"""
-utils fgt get
-"""
-
-from .version_ import version
-
-__all__ = ["version"]
