+++ fotoobo/utils/fmg/__init__.py
@@ -3,7 +3,6 @@ utils fmg
 """
 
 from . import get
-from .assign_ import assign
-from .set_ import set  # pylint: disable=redefined-builtin
+from .main import assign, set  # pylint: disable=redefined-builtin
 
 __all__ = ["assign", "get", "set"]
