--- fotoobo/utils/fmg/get/__init__.py
@@ -1,10 +1,0 @@
-"""
-utils fmg get
-"""
-
-from .adoms_ import adoms
-from .devices_ import devices
-from .policy_ import policy
-from .version_ import version
-
-__all__ = ["adoms", "devices", "policy", "version"]
